Harvest API v2 get request wrapper

Submits get requests to the Harvest API v2. The function is a wrapper around the httr::GET function and returns the Harvest table.

get_table(
  table = NULL,
  user = harvestR::get_harvest_account(),
  key = harvestR::get_harvest_pat(),
  email = "",
  query = NULL,
  ...,
  verbose = lifecycle::deprecated(),
  strategy = lifecycle::deprecated()
)

Arguments

table
Character table name to be submitted to the Harvest API v2. Refer to Harvest API v2.

user
Character account number (e.g. user = '12345') for authentication. Register at Harvest Developers. Defaults to `get_harvest_account()`

key
Character key (e.g. key = 'Bearer <secret key>') for API authentication. Register at Harvest Developers. Defaults to `get_harvest_pat()`

email
Character email address for API call. (Optional)

...
Additional arguments, some are forwarded to other functions - see details.

Value

dataframe with contents of requested Harvest table

Details

Description of ... parameters

auto_retry Logical; control whether to retry GET requests upon failure. TRUE results 2 retries (3 attempts total). FALSE results in no retries (1 attempt total).

quiet Logical; controls messaging. TRUE will result in displaying only errors in the console. FALSE will results in displaying messages, warnings, and errors. Depending on execution strategy, messages from httr::GET may not be surfaced to the console due to multiple R sessions running and stderror relaying issues. See Future Package Output for me details.

plan_options List; list of parameters passed to future::plan. The strategy parameter is required to implement parallel requests, see Future Package Overview for more details.

References

Harvest API v2 Harvest Developers Page Future Pacakge Overview Future Package Output

See also

Examples

if (FALSE) {
create_harvest_creds()
set_harvest_creds()
get_table(table = 'projects')
get_table(table = 'time_entries',
          email = 'your_email',
          query = list(from = '2018-01-01', to = '2018-12-31'),
          plan_options = list(strategy = "multicore"),
          auto_retry = T,
          quiet = T)
}